Move `TextDirection` to `model.ts`
