+++ src/vs/monaco.d.ts
@@ -1678,12 +1678,6 @@ declare namespace monaco.editor {
 		sectionHeaderText?: string | null;
 	}
 
-	export enum TextDirection {
-		LTR = 0,
-
-		RTL = 1,
-	}
-
 	/**
 	 * Options for a model decoration.
 	 */
@@ -1818,11 +1812,18 @@ declare namespace monaco.editor {
 		 * If set, text will be injected in the view before the range.
 		 */
 		before?: InjectedTextOptions | null;
-
 		/**
 		 * The text direction of the decoration.
 		 */
 		textDirection?: TextDirection | null;
+	}
+
+	/**
+	 * Text Direction for a decoration.
+	 */
+	export enum TextDirection {
+		LTR = 0,
+		RTL = 1
 	}
 
 	/**
